backend/requirements.txt
# /backend/requirements.txt

Flask
google-generativeai
python-dotenv
requests
markdown2
Flask-Cors
gunicorn
